Specify data type in queries for constant values.
svn commit r4248

+++ Admin/AdminListDependency.php
@@ -210,7 +210,7 @@ class AdminListDependency extends AdminDependency
 		}
 
 		$sql = sprintf('select %s as id, %s as parent, %s as title,
-				%s as status_level from %s',
+				%s::integer as status_level from %s',
 			$id_field->name, $parent_value, $title_field->name, $status_level,
 			$table);
